billing calculations refactored
